modified:   Estilos.css ya se pude ver el estili en el boton de agrefar carrito modified:   Productos.js
modified:   actualizar.js
new file:   adminPanel.html
new file:   administacion.js
implementacion del la pagina como l profe quiso
agrege validacion a datos en actualizar y agregar producto

+++ eliminarProducto.js
@@ -11,7 +11,13 @@ function (evt){
     console.log("entro event")
 evt.preventDefault();
 const idProducto = document.getElementById("id").value;
-eliminarProducto(idProducto);
+if (!isNaN(idProducto)) {
+    eliminarProducto(idProducto);
+}
+else{
+   alert("id  tiene que ser un numero")
+}
+
 });
 
 function eliminarProducto(_idProducto) {
